+++ src/events/misc.h
@@ -9,5 +9,6 @@ void event_nicknameInUse                  (struct irc_message_compo *);
 void event_channel_forward                (struct irc_message_compo *);
 void event_channelModeIs                  (struct irc_message_compo *);
 void event_channelCreatedWhen             (struct irc_message_compo *);
+void event_userModeIs                     (struct irc_message_compo *);
 
 #endif