+++ src/statusbar.h
@@ -1,6 +1,8 @@
 #ifndef STATUSBAR_H
 #define STATUSBAR_H
 
+extern char g_user_modes[100];
+
 void	statusbar_deinit(void);
 void	statusbar_hide(void);
 void	statusbar_init(void);